refactor(web): split Assets.tsx — extract section bars and tx-picker view
